third_party/blink/web_tests/wpt_internal/display-lock/rendersubtree/measure-updated-layout.html

<!doctype HTML>
<html>
<meta charset="utf8">
<title>Display Locking: measure updated layout</title>
<link rel="author" title="Vladimir Levin" href="mailto:vmpstr@chromium.org">
<link rel="help" href="https://github.com/WICG/display-locking">
<script src="resources/utils.js"></script>

<style>
#container {
  background: lightgreen;
}
#sizer {
  width: 100px;
  height: 100px;
}
.child {
  width: 20px;
  height: 20%;
  background: cyan;
}
#spacer {
  width: 150px;
  height: 150px;
  background: green;
}
</style>

<div id="parent"></div>
<div id="spacer"></div>

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
async_test((t) => {
  function createChild(id) {
    const child = document.createElement("div");
    child.classList = "child";
    child.id = id;
    return child;
  }

  function construct(container) {
    const sizer = document.createElement("div");
    sizer.id = "sizer";
    container.appendChild(sizer);
    sizer.appendChild(createChild("0"));
    sizer.appendChild(createChild("1"));
    sizer.appendChild(createChild("2"));
  }

  function measureInUpdate() {
    t.step(() => {
      // Ensure children are laid out; this forces a layout.
      assert_equals(document.getElementById("0").offsetTop, 0, "0 in update");
      assert_equals(document.getElementById("1").offsetTop, 20, "1 in update");
      assert_equals(document.getElementById("2").offsetTop, 40, "2 in update");
      // Both parent should be 0 height, since it's locked. Both parent and spacers
      // should have 8 offsetTop.
      assert_equals(document.getElementById("parent").offsetTop, 8, "parent in update");
      assert_equals(document.getElementById("spacer").offsetTop, 8, "spacer in update");
    });
  }

  function measureInCommit() {
    t.step(() => {
      // Ensure children are still laid out.
      assert_equals(document.getElementById("0").offsetTop, 0, "0 in commit");
      assert_equals(document.getElementById("1").offsetTop, 20, "1 in commit");
      assert_equals(document.getElementById("2").offsetTop, 40, "2 in commit");
      // Now the parent should encompass a container, so spacer is pushed down.
      assert_equals(document.getElementById("parent").offsetTop, 8, "parent in commit");
      assert_equals(document.getElementById("spacer").offsetTop, 108, "spacer in commit");
    });
  }

  async function runTest() {
    const container = document.createElement("div");
    container.id = "container";
    await setInvisible(container);

    document.getElementById("parent").appendChild(container);
    construct(container);

    container.updateRendering().then(() => {
      measureInUpdate();
      setVisible(container).then(() => {
        measureInCommit();
        t.done();
      });
    });
  }

  window.onload = function() {
    requestAnimationFrame(() => requestAnimationFrame(runTest));
  };
}, "Measure Updated Layout");
</script>
</html>
